Working network growth and weight distribution updates
Successfully simulates 1000 new vertices

+++ include/NetworkComponents.hpp
@@ -10,15 +10,15 @@ using std::vector, std::array, std::cout;
 
 class Vertex{
     public:
-        Vertex(int , double , int ); // constructor
-        double calculateWeight(); // calculates the weight of the node, in the case of Bianconi-Barabasi model this is degree * fitness
-        double getFitness(); // returns the fitness of the vertex
+        Vertex(int , long double , int ); // constructor
+        long double calculateWeight(); // calculates the weight of the node, in the case of Bianconi-Barabasi model this is degree * fitness
+        long double getFitness(); // returns the fitness of the vertex
         void increaseDegree(); // increments the degree of the vertex by one
         int getName(); // returns the name of the vertex
         int getDegree(); // returns the degree of the vertex
     private:
         const int name; // vertex is named with an int which denotes the "time" it was created, meaning that an older vertex will have a lower integer name
-        const double fitness; // fitness of the vertex, usually chosen randomly according to some distribution
+        const long double fitness; // fitness of the vertex, usually chosen randomly according to some distribution
         int degree; // degree of the vertex
 };
 std::ostream& operator<<(std::ostream&, Vertex& );
